refactor: updated project structure
