+++ src/App.js
@@ -2,10 +2,10 @@ import React from "react";
 
 import { BrowserRouter as Router, Switch, Route } from "react-router-dom";
 
-import { AuthProvider } from "../contexts/AuthContext";
+import { AuthProvider } from "./contexts/AuthContext";
 
-import Chats from "./Chats";
-import Login from "./Login";
+import Chats from "./components/Chats";
+import Login from "./pages/Login";
 
 function App() {
   return (
