--- src/components/Login.js
@@ -1,26 +1,0 @@
-import React from "react";
-import { GoogleOutlined, FacebookOutlined } from "@ant-design/icons";
-import firebase from "firebase/app";
-
-import { auth } from "../firebase";
-
-const Login = () => {
-  return (
-    <div id="login-page">
-      <div id="login-card">
-        <h2>Welcome to the Chat</h2>
-        <div
-          className="login-button google"
-          onClick={() =>
-            auth.signInWithRedirect(new firebase.auth.GoogleAuthProvider())
-          }
-        >
-          <GoogleOutlined /> Sign in with Google
-        </div>
-        <br /> <br />
-      </div>
-    </div>
-  );
-};
-
-export default Login;
